[fix] 테스트 에러 수정

+++ dutchiepay/src/app/hooks/useDeleteReview.jsx
@@ -1,18 +1,19 @@
-// hooks/useDeleteReview.js
-
 import axios from 'axios';
-import { useCallback } from 'react';
+import { useCallback, useRef } from 'react';
 import useReissueToken from './useReissueToken';
 import { useSelector } from 'react-redux';
 
 const useDeleteReview = () => {
   const access = useSelector((state) => state.login.access);
   const { refreshAccessToken } = useReissueToken();
+  const hasFetched = useRef(false);
   const deleteReview = useCallback(
     async (reviewId) => {
       const confirmed = confirm('작성한 리뷰를 삭제하시겠습니까?');
-      if (!confirmed) return false; // 사용자가 취소한 경우 false 반환
+      if (!confirmed) return false;
+      if (hasFetched.current) return;
 
+      hasFetched.current = true;
       try {
         await axios.delete(
           `${process.env.NEXT_PUBLIC_BASE_URL}/profile/reviews?reviewId=${reviewId}`,
@@ -23,10 +24,10 @@ const useDeleteReview = () => {
           }
         );
         alert('정상적으로 삭제되었습니다.');
-        return true; // 삭제 성공
+        return true;
       } catch (error) {
         if (error.response.data.message === '액세스 토큰이 만료되었습니다.') {
-          /*const reissueResponse = await refreshAccessToken();
+          const reissueResponse = await refreshAccessToken();
           if (reissueResponse.success) {
             await deleteReview(reviewId);
           } else {
@@ -34,11 +35,11 @@ const useDeleteReview = () => {
               reissueResponse.message ||
                 '오류가 발생했습니다. 다시 시도해주세요.'
             );
-          }*/
+          }
         } else {
           alert('오류가 발생했습니다. 다시 시도해주세요.');
         }
-        return false; // 삭제 실패
+        return false;
       }
     },
     [access]
